refactor(persistence): reuse config path authority
