+++ src/ouroboros/cli/commands/tui.py
@@ -17,8 +17,8 @@ from typing import Annotated
 import typer
 
 from ouroboros.cli.formatters.panels import print_error, print_info, print_success
+from ouroboros.config.models import resolve_event_store_path
 from ouroboros.persistence.event_store import EventStore
-from ouroboros.persistence.paths import resolve_event_store_path
 
 app = typer.Typer(
     name="tui",
